+++ libs/store/src/lib/operators/index.ts
@@ -1,2 +1,1 @@
 export { untilDestroy, destroy$ as ɵdestroy$ } from './untilDestroy';
-export * from './pluck';